git commit -am "refactor: reintroduce Recorder
* rename AudioRecording -> Recording
* make Provider::recognize take bytes
* add recognize_state prop to Recording
