+++ src/main.rs
@@ -25,7 +25,6 @@
 mod about;
 mod application;
 mod audio_device;
-mod audio_recording;
 mod clock_time;
 mod config;
 mod core;
