+++ src/recognizer/provider/mod.rs
@@ -8,12 +8,12 @@ use async_trait::async_trait;
 use std::{fmt, time::Duration};
 
 pub use self::settings::{ProviderSettings, ProviderType, TestProviderMode};
-use crate::{audio_recording::AudioRecording, model::Song};
+use crate::model::Song;
 
 #[async_trait(?Send)]
 pub trait Provider: fmt::Debug {
-    /// Recognize a song from a recording
-    async fn recognize(&self, recording: &AudioRecording) -> Result<Song>;
+    /// Recognize a song from bytes
+    async fn recognize(&self, bytes: &[u8]) -> Result<Song>;
 
     /// How long to record the audio
     fn listen_duration(&self) -> Duration;
@@ -26,11 +26,7 @@ pub trait Provider: fmt::Debug {
 
 #[async_trait(?Send)]
 pub trait TestProvider {
-    async fn recognize_impl(
-        &self,
-        recording: &AudioRecording,
-        mode: TestProviderMode,
-    ) -> Result<Song>;
+    async fn recognize_impl(&self, bytes: &[u8], mode: TestProviderMode) -> Result<Song>;
 }
 
 #[async_trait(?Send)]
@@ -38,12 +34,12 @@ impl<T> Provider for T
 where
     T: TestProvider + fmt::Debug,
 {
-    async fn recognize(&self, recording: &AudioRecording) -> Result<Song> {
+    async fn recognize(&self, bytes: &[u8]) -> Result<Song> {
         let duration = ProviderSettings::lock().test_recognize_duration;
         glib::timeout_future(duration).await;
 
         let mode = ProviderSettings::lock().test_mode;
-        self.recognize_impl(recording, mode).await
+        self.recognize_impl(bytes, mode).await
     }
 
     fn listen_duration(&self) -> Duration {
